requirements.txt
torch
torchvision
fastapi
uvicorn
PILLOW
openai
streamlit
requests